close #12 added function to get new images from camera or gallery integrated with occurrence add
