--- src/pages/get-image/get-image.html
@@ -1,21 +1,0 @@
-<!--
-  Generated template for the GetImagePage page.
-
-  See http://ionicframework.com/docs/components/#navigation for more info on
-  Ionic pages and navigation.
--->
-<ion-header>
-
-  <ion-navbar>
-    <ion-title>Get Image</ion-title>
-  </ion-navbar>
-
-</ion-header>
-
-<ion-content padding>
-<ion-content padding>
-  <ion-avatar (click)="changePicture()">
-    <img [src]="chosenPicture || placeholder" onerror="this.src='assets/img/avatar/girl-avatar.png'" />
-    <button ion-button color="danger">Change image</button>
-  </ion-avatar>
-</ion-content>
