--- src/pages/get-image/get-image.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { IonicPageModule } from 'ionic-angular';
-import { GetImagePage } from './get-image';
-
-@NgModule({
-  declarations: [
-    GetImagePage,
-  ],
-  imports: [
-    IonicPageModule.forChild(GetImagePage),
-  ],
-  exports: [
-    GetImagePage
-  ]
-})
-export class GetImagePageModule {}
