--- src/pages/get-image/get-image.scss
@@ -1,12 +1,0 @@
-page-get-image {
-  ion-content {
-    text-align: center;
-  }
-  ion-avatar img,
-  ion-thumbnail img {
-    display: block;
-    margin: 0 auto;
-    width: 25vh;
-    height: auto;
-  }
-}
