--- src/pages/get-image/get-image.ts
@@ -1,78 +1,0 @@
-import { CameraProvider } from './../../providers/camera/camera.provider';
-import { Component } from '@angular/core';
-import { NavController, Platform, ActionSheetController, LoadingController, IonicPage } from 'ionic-angular';
-
-@IonicPage()
-@Component({
-  selector: 'page-get-image',
-  templateUrl: 'get-image.html'
-})
-export class GetImagePage {
-
-  placeholder = 'assets/img/avatar/girl-avatar.png';
-  chosenPicture: any;
-
-  constructor(public navCtrl: NavController, public actionsheetCtrl: ActionSheetController,
-    public cameraProvider: CameraProvider, public platform: Platform, public loadingCtrl: LoadingController) {
-  }
-
-  changePicture() {
-
-    let actionsheet = this.actionsheetCtrl.create({
-      title: 'upload picture',
-      buttons: [
-        {
-          text: 'camera',
-          icon: !this.platform.is('ios') ? 'camera' : null,
-          handler: () => {
-            this.takePicture();
-          }
-        },
-        {
-          text: !this.platform.is('ios') ? 'gallery' : 'camera roll',
-          icon: !this.platform.is('ios') ? 'image' : null,
-          handler: () => {
-            this.getPicture();
-          }
-        },
-        {
-          text: 'cancel',
-          icon: !this.platform.is('ios') ? 'close' : null,
-          role: 'destructive',
-          handler: () => {
-            console.log('the user has cancelled the interaction.');
-          }
-        }
-      ]
-    });
-    return actionsheet.present();
-  }
-
-  takePicture() {
-    let loading = this.loadingCtrl.create();
-
-    loading.present();
-    return this.cameraProvider.getPictureFromCamera().then(picture => {
-      if (picture) {
-        this.chosenPicture = picture;
-      }
-      loading.dismiss();
-    }, error => {
-      alert(error);
-    });
-  }
-
-  getPicture() {
-    let loading = this.loadingCtrl.create();
-
-    loading.present();
-    return this.cameraProvider.getPictureFromPhotoLibrary().then(picture => {
-      if (picture) {
-        this.chosenPicture = picture;
-      }
-      loading.dismiss();
-    }, error => {
-      alert(error);
-    });
-  }
-}
