+++ src/pages/occurrence-add/occurrence-add.html
@@ -17,13 +17,16 @@
 
 
 <ion-content padding>
-  <ion-item>
-    <ion-label floating>Descrição</ion-label>
-    <ion-input type="text" [(ngModel)]="occurrenceItem.description"></ion-input>
-  </ion-item>
-  <ion-item>
-    <button ion-item (click)="itemTapped($event)">
-      info
-    </button>
-  </ion-item>
+  <ion-list>
+    <ion-item>
+      <ion-label floating>Descrição</ion-label>
+      <ion-input type="text" [(ngModel)]="occurrenceItem.description"></ion-input>
+    </ion-item>
+    <ion-item>
+      <ion-avatar (click)="changePicture()">
+        <img [src]="chosenPicture || placeholder" onerror="this.src='assets/img/avatar/girl-avatar.png'" />
+        <button ion-button color="danger">Capturar Imagem</button>
+      </ion-avatar>
+    </ion-item>
+  </ion-list>
 </ion-content>
